requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
pytest==7.4.0
python-dotenv==1.0.0
gunicorn==21.2.0
SQLAlchemy==2.0.20
Werkzeug==2.3.7
waitress==2.1.2
pytest-cov==4.1.0
marshmallow==3.19.0